+++ src/main/java/io/github/manuelarte/spring/queryparameter/query/QueryCriteria.java
@@ -2,14 +2,20 @@ package io.github.manuelarte.spring.queryparameter.query;
 
 import com.google.common.base.Preconditions;
 import io.github.manuelarte.spring.queryparameter.operators.Operator;
+import java.util.Iterator;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.Spliterator;
+import java.util.function.Consumer;
+import java.util.stream.Stream;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import lombok.val;
 
 @lombok.EqualsAndHashCode
+@lombok.ToString
 @lombok.Builder(toBuilder = true)
-public final class QueryCriteria {
+public final class QueryCriteria implements Iterable<QueryCriterion<Object>> {
 
   private final QueryCriterion<?> criterion;
   private final OtherCriteria other;
@@ -40,6 +46,59 @@ public final class QueryCriteria {
 
   public Optional<OtherCriteria> getOther() {
     return Optional.ofNullable(other);
+  }
+
+  @Override
+  public Iterator<QueryCriterion<Object>> iterator() {
+    return new QueryCriteriaIterator(this);
+  }
+
+  @Override
+  public void forEach(Consumer<? super QueryCriterion<Object>> action) {
+    new QueryCriteriaIterator(this).forEachRemaining(action);
+  }
+
+  public static class QueryCriteriaIterator implements Iterator<QueryCriterion<Object>> {
+
+    private final QueryCriteria queryCriteria;
+    private int position = 0;
+
+    public QueryCriteriaIterator(final QueryCriteria queryCriteria) {
+      this.queryCriteria = queryCriteria;
+    }
+
+    @Override
+    public boolean hasNext() {
+      return getNext(position).isPresent();
+    }
+
+    @Override
+    public QueryCriterion<Object> next() {
+      val toReturn = getNext(position).get();
+      position++;
+      return toReturn;
+    }
+
+    private Optional<QueryCriterion<Object>> getNext(final int index) {
+      int accIndex = 0;
+      if (index == 0) {
+        return Optional.of(queryCriteria.getCriterion());
+      } else {
+        QueryCriteria acc = this.queryCriteria;
+        Optional<OtherCriteria> otherCriteria = Optional.empty();
+        while (accIndex < index) {
+          if (acc.getOther().isPresent()) {
+            otherCriteria = acc.getOther();
+            acc = otherCriteria.get().getCriteria();
+            accIndex++;
+          } else {
+            return Optional.empty();
+          }
+        }
+        return otherCriteria.map(it -> it.getCriteria().getCriterion());
+      }
+    }
+
   }
 
   public static class QueryCriteriaBuilder {