+++ src/test/java/io/github/manuelarte/spring/queryparameter/query/QueryCriteriaTest.java
@@ -3,6 +3,8 @@ package io.github.manuelarte.spring.queryparameter.query;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import io.github.manuelarte.spring.queryparameter.operators.EqualsOperator;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.StreamSupport;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import org.junit.jupiter.api.Test;
 
@@ -22,13 +24,13 @@ public class QueryCriteriaTest {
   @Test
   public void testAddInBuilder() {
     final QueryCriteria actual = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
-        .and(new QueryCriterion("key2", new EqualsOperator(), "value2")).build();
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
+        .and(new QueryCriterion<>("key2", new EqualsOperator(), "value2")).build();
 
     final QueryCriteria expected = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
         .other(new OtherCriteria(BooleanOperator.AND,
-            new QueryCriteria(new QueryCriterion("key2",
+            new QueryCriteria(new QueryCriterion<>("key2",
                 new EqualsOperator(), "value2")))).build();
     assertEquals(expected, actual);
   }
@@ -36,19 +38,19 @@ public class QueryCriteriaTest {
   @Test
   public void testTwoAddInBuilder() {
     final QueryCriteria actual = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
-        .and(new QueryCriterion("key2", new EqualsOperator(), "value2"))
-        .and(new QueryCriterion("key3", new EqualsOperator(), "value3"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
+        .and(new QueryCriterion<>("key2", new EqualsOperator(), "value2"))
+        .and(new QueryCriterion<>("key3", new EqualsOperator(), "value3"))
         .build();
 
     final QueryCriteria expected = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
         .other(
             new OtherCriteria(BooleanOperator.AND,
                 new QueryCriteria(
-                    new QueryCriterion("key2", new EqualsOperator(), "value2"),
+                    new QueryCriterion<>("key2", new EqualsOperator(), "value2"),
                     new OtherCriteria(BooleanOperator.AND, new QueryCriteria(
-                        new QueryCriterion("key3", new EqualsOperator(), "value3"))))
+                        new QueryCriterion<>("key3", new EqualsOperator(), "value3"))))
             )
         )
         .build();
@@ -58,13 +60,13 @@ public class QueryCriteriaTest {
   @Test
   public void testOrInBuilder() {
     final QueryCriteria actual = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
-        .or(new QueryCriterion("key2", new EqualsOperator(), "value2")).build();
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
+        .or(new QueryCriterion<>("key2", new EqualsOperator(), "value2")).build();
 
     final QueryCriteria expected = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
         .other(new OtherCriteria(BooleanOperator.OR,
-            new QueryCriteria(new QueryCriterion("key2",
+            new QueryCriteria(new QueryCriterion<>("key2",
                 new EqualsOperator(), "value2")))).build();
     assertEquals(expected, actual);
   }
@@ -72,19 +74,19 @@ public class QueryCriteriaTest {
   @Test
   public void testTwoOrInBuilder() {
     final QueryCriteria actual = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
-        .or(new QueryCriterion("key2", new EqualsOperator(), "value2"))
-        .or(new QueryCriterion("key3", new EqualsOperator(), "value3"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
+        .or(new QueryCriterion<>("key2", new EqualsOperator(), "value2"))
+        .or(new QueryCriterion<>("key3", new EqualsOperator(), "value3"))
         .build();
 
     final QueryCriteria expected = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
         .other(
             new OtherCriteria(BooleanOperator.OR,
                 new QueryCriteria(
-                    new QueryCriterion("key2", new EqualsOperator(), "value2"),
+                    new QueryCriterion<>("key2", new EqualsOperator(), "value2"),
                     new OtherCriteria(BooleanOperator.OR, new QueryCriteria(
-                        new QueryCriterion("key3", new EqualsOperator(), "value3"))))
+                        new QueryCriterion<>("key3", new EqualsOperator(), "value3"))))
             )
         )
         .build();
@@ -94,19 +96,19 @@ public class QueryCriteriaTest {
   @Test
   public void testAndOrInBuilder() {
     final QueryCriteria actual = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
-        .and(new QueryCriterion("key2", new EqualsOperator(), "value2"))
-        .or(new QueryCriterion("key3", new EqualsOperator(), "value3"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
+        .and(new QueryCriterion<>("key2", new EqualsOperator(), "value2"))
+        .or(new QueryCriterion<>("key3", new EqualsOperator(), "value3"))
         .build();
 
     final QueryCriteria expected = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
         .other(
             new OtherCriteria(BooleanOperator.AND,
                 new QueryCriteria(
-                    new QueryCriterion("key2", new EqualsOperator(), "value2"),
+                    new QueryCriterion<>("key2", new EqualsOperator(), "value2"),
                     new OtherCriteria(BooleanOperator.OR, new QueryCriteria(
-                        new QueryCriterion("key3", new EqualsOperator(), "value3"))))
+                        new QueryCriterion<>("key3", new EqualsOperator(), "value3"))))
             )
         )
         .build();
@@ -116,23 +118,36 @@ public class QueryCriteriaTest {
   @Test
   public void testOrAndInBuilder() {
     final QueryCriteria actual = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
-        .or(new QueryCriterion("key2", new EqualsOperator(), "value2"))
-        .and(new QueryCriterion("key3", new EqualsOperator(), "value3"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
+        .or(new QueryCriterion<>("key2", new EqualsOperator(), "value2"))
+        .and(new QueryCriterion<>("key3", new EqualsOperator(), "value3"))
         .build();
 
     final QueryCriteria expected = QueryCriteria.builder()
-        .criterion(new QueryCriterion("key", new EqualsOperator(), "value"))
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
         .other(
             new OtherCriteria(BooleanOperator.OR,
                 new QueryCriteria(
-                    new QueryCriterion("key2", new EqualsOperator(), "value2"),
+                    new QueryCriterion<>("key2", new EqualsOperator(), "value2"),
                     new OtherCriteria(BooleanOperator.AND, new QueryCriteria(
-                        new QueryCriterion("key3", new EqualsOperator(), "value3"))))
+                        new QueryCriterion<>("key3", new EqualsOperator(), "value3"))))
             )
         )
         .build();
     assertEquals(expected, actual);
+  }
+
+  @Test
+  void testIterableWithThreeCriterion() {
+    final QueryCriteria actual = QueryCriteria.builder()
+        .criterion(new QueryCriterion<>("key", new EqualsOperator(), "value"))
+        .and(new QueryCriterion<>("key2", new EqualsOperator(), "value2"))
+        .or(new QueryCriterion<>("key3", new EqualsOperator(), "value3"))
+        .build();
+
+    final AtomicInteger atomicInteger = new AtomicInteger(0);
+    actual.forEach(it -> atomicInteger.addAndGet(1));
+    assertEquals(3, atomicInteger.get());
   }
 
 }