+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-### Changed
+- Added query parameter constraints to add in entities NotNull and Null
 
+### Changed
 
 ### Removed